# swagger-spec.json
{"openapi":"3.0.0","paths":{"/management/register":{"post":{"operationId":"ManagementAuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"user created"},"400":{"description":"Request body not match"},"409":{"description":"Email already exist"}},"tags":["Management"],"security":[{"bearer":[]}]}},"/management/login":{"post":{"operationId":"ManagementAuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Success Login.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDTO"}}}}},"tags":["Management"],"security":[{"bearer":[]}]}},"/management/profile":{"get":{"operationId":"ManagementAuthController_getManagementProfile","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Management"}}}}}},"tags":["Management"],"security":[{"bearer":[]}]}},"/management/branch":{"post":{"operationId":"ManagementAuthController_createBranch","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBranchDto"}}}},"responses":{"201":{"description":""}},"tags":["Management"],"security":[{"bearer":[]}]},"get":{"operationId":"ManagementAuthController_getAllBranch","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Branch"}}}}}},"tags":["Management"],"security":[{"bearer":[]}]}},"/management/operating-area":{"get":{"operationId":"ManagementAuthController_getOperatingArea","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OperatingArea"}}}}}},"tags":["Management"],"security":[{"bearer":[]}]},"post":{"operationId":"ManagementAuthController_createOperatingArea","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOperatingAreaDto"}}}},"responses":{"201":{"description":""}},"tags":["Management"],"security":[{"bearer":[]}]}},"/management/operating-area/{id}":{"put":{"operationId":"ManagementAuthController_updateOperatingArea","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOperatingAreaDto"}}}},"responses":{"200":{"description":""}},"tags":["Management"],"security":[{"bearer":[]}]},"delete":{"operationId":"ManagementAuthController_deleteOperatingArea","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Management"],"security":[{"bearer":[]}]}},"/branch/reset-password":{"post":{"operationId":"BranchAuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminResetPasswordDTO"}}}},"responses":{"201":{"description":"Success change Password"},"400":{"description":"Token Invalid OR password not strong enough"}},"tags":["Branch"],"security":[{"bearer":[]}]}},"/branch/login":{"post":{"operationId":"BranchAuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminLoginDto"}}}},"responses":{"200":{"description":"Success Login.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminLoginResponseDTO"}}}}},"tags":["Branch"],"security":[{"bearer":[]}]}},"/branch/profile":{"get":{"operationId":"BranchAuthController_getBranchProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["Branch"],"security":[{"bearer":[]}]}}},"info":{"title":"KTSM User Auth Service","description":"API for User data","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","default":"test@gmail.com"},"password":{"type":"string","default":"Password123!"}},"required":["email","password"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","default":"test@gmail.com"},"password":{"type":"string","default":"Password123!"}},"required":["email","password"]},"Management":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","email","createdAt","updatedAt"]},"DataDTO":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/Management"},"token":{"type":"string"}},"required":["user","token"]},"LoginResponseDTO":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/DataDTO"},"success":{"type":"boolean"}},"required":["data","success"]},"CreateBranchDto":{"type":"object","properties":{"email":{"type":"string","default":"test@gmail.com"},"operatingAreaId":{"type":"string","default":"0557d6ef-bc10-4a9a-967f-0520c3abeb55","description":"You can get operating area id, from get opreating area"}},"required":["email","operatingAreaId"]},"OperatingArea":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["id","name","createdAt","updatedAt","deletedAt"]},"Branch":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"operatingArea":{"$ref":"#/components/schemas/OperatingArea"},"password":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","email","operatingArea","password","createdAt","updatedAt"]},"CreateOperatingAreaDto":{"type":"object","properties":{"name":{"type":"string","default":"Jakarta Selatan"}},"required":["name"]},"AdminResetPasswordDTO":{"type":"object","properties":{"password":{"type":"string"}},"required":["password"]},"AdminLoginDto":{"type":"object","properties":{"email":{"type":"string","default":"superadmin@gmail.com"},"password":{"type":"string","default":"SuperPassword123!"}},"required":["email","password"]},"AdminLoginResponseDTO":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/DataDTO"},"success":{"type":"boolean"}},"required":["data","success"]}}}}